pass only certificate filepath through rest middleware.

+++ tests/fixtures.py
@@ -13,16 +13,15 @@ from sqlalchemy.engine import create_engine
 from nucypher.blockchain.eth.deployers import PolicyManagerDeployer, NucypherTokenDeployer, MinerEscrowDeployer
 from nucypher.blockchain.eth.interfaces import BlockchainDeployerInterface
 from nucypher.blockchain.eth.registry import TemporaryEthereumContractRegistry
-
 from nucypher.blockchain.eth.sol.compile import SolidityCompiler
-from nucypher.characters.lawful import Alice, Bob
-from nucypher.config.characters import UrsulaConfiguration
-from nucypher.config.constants import TEST_CONTRACTS_DIR, BASE_DIR
+from nucypher.characters.lawful import Bob
+from nucypher.config.characters import UrsulaConfiguration, AliceConfiguration
+from nucypher.config.constants import BASE_DIR
+from nucypher.config.constants import TEST_CONTRACTS_DIR
 from nucypher.data_sources import DataSource
 from nucypher.keystore import keystore
 from nucypher.keystore.db import Base
 from nucypher.keystore.keypairs import SigningKeypair
-
 from nucypher.utilities.sandbox.blockchain import TesterBlockchain, token_airdrop
 from nucypher.utilities.sandbox.constants import (TEST_URSULA_STARTING_PORT,
                                                   DEFAULT_NUMBER_OF_URSULAS_IN_DEVELOPMENT_NETWORK,
@@ -44,6 +43,16 @@ def tempfile_path():
     os.remove(path)
 
 
+@pytest.fixture(scope="module", autouse=True)
+def temp_config_root():
+    """
+    User is responsible for closing the file given at the path.
+    """
+    temp_dir = tempfile.TemporaryDirectory(prefix='nucypher-tmp-')
+    yield temp_dir.name
+    temp_dir.cleanup()
+
+
 @pytest.fixture(scope="module")
 def test_keystore():
     engine = create_engine('sqlite:///:memory:')
@@ -56,9 +65,10 @@ def test_keystore():
 # Configuration
 #
 
-@pytest.fixture(scope="session")
-def temporary_ursula_config():
-    ursula_config = UrsulaConfiguration(temp=True)
+@pytest.fixture(scope="module")
+def temporary_ursula_config(temp_config_root):
+    ursula_config = UrsulaConfiguration(temp=True,
+                                        config_root=temp_config_root)
     yield ursula_config
 
 
@@ -89,6 +99,7 @@ def enacted_federated_policy(idle_federated_policy, federated_ursulas):
                                             deposit=deposit,
                                             expiration=contract_end_datetime,
                                             handpicked_ursulas=federated_ursulas)
+
     responses = idle_federated_policy.enact(network_middleware)  # REST call happens here, as does population of TreasureMap.
 
     return idle_federated_policy
@@ -123,26 +134,41 @@ def enacted_blockchain_policy(idle_blockchain_policy, blockchain_ursulas):
 #
 
 @pytest.fixture(scope="module")
-def alice(federated_ursulas):
-    alice = Alice(network_middleware=MockRestMiddleware(),
-                  known_nodes=federated_ursulas,
-                  federated_only=True,
-                  abort_on_learning_error=True)
+def alice_federated_test_config(federated_ursulas, temp_config_root):
+    config = AliceConfiguration(temp=True,
+                                config_root=temp_dir,
+                                network_middleware=MockRestMiddleware(),
+                                known_nodes=federated_ursulas,
+                                federated_only=True,
+                                abort_on_learning_error=True)
+    yield config
+
+
+@pytest.fixture(scope="module")
+def alice_blockchain_test_config(blockchain_ursulas, three_agents, temp_config_root):
+    token_agent, miner_agent, policy_agent = three_agents
+    etherbase, alice_address, bob_address, *everyone_else = token_agent.blockchain.interface.w3.eth.accounts
+
+    config = AliceConfiguration(temp=True,
+                                config_root=temp_dir,
+                                network_middleware=MockRestMiddleware(),
+                                policy_agent=policy_agent,
+                                known_nodes=blockchain_ursulas,
+                                abort_on_learning_error=True,
+                                checksum_address=alice_address)
+
+    yield config
+
+
+@pytest.fixture(scope="module")
+def alice(federated_ursulas, alice_federated_test_config):
+    alice = alice_federated_test_config.produce()
     return alice
 
 
 @pytest.fixture(scope="module")
-def blockchain_alice(blockchain_ursulas, three_agents):
-    token_agent, miner_agent, policy_agent = three_agents
-    etherbase, alice_address, bob_address, *everyone_else = token_agent.blockchain.interface.w3.eth.accounts
-
-    alice = Alice(network_middleware=MockRestMiddleware(),
-                  policy_agent=policy_agent,
-                  known_nodes=blockchain_ursulas,
-                  abort_on_learning_error=True,
-                  checksum_address=alice_address)
-    # alice.recruit = lambda *args, **kwargs: [u._ether_address for u in ursulas]
-
+def blockchain_alice(blockchain_ursulas, alice_blockchain_test_config):
+    alice = alice_blockchain_test_config.produce()
     return alice
 
 
@@ -169,22 +195,24 @@ def capsule_side_channel(enacted_federated_policy):
 #
 
 @pytest.fixture(scope="module")
-def federated_ursulas():
+def federated_ursulas(temp_config_root):
+    _ursulas = None
     try:
-        pathlib.Path(CERT_DIR_FOR_TEST_FIXTURES).mkdir(parents=True, exist_ok=True)
         _ursulas = make_federated_ursulas(quantity=DEFAULT_NUMBER_OF_URSULAS_IN_DEVELOPMENT_NETWORK,
-                                          certificate_dir=CERT_DIR_FOR_TEST_FIXTURES)
+                                          config_root=temp_config_root)
         yield _ursulas
+    except Exception as e:
+        raise
     finally:
-        # Remove the DBs that have been sprayed hither and yon.
-        with contextlib.suppress(FileNotFoundError):
-            for port, ursula in enumerate(_ursulas, start=TEST_URSULA_STARTING_PORT):
-                os.remove("test-{}".format(port))
-        shutil.rmtree(CERT_DIR_FOR_TEST_FIXTURES)
+        if _ursulas:
+            # Remove the DBs that have been sprayed hither and yon.
+            with contextlib.suppress(FileNotFoundError):
+                for port, ursula in enumerate(_ursulas, start=TEST_URSULA_STARTING_PORT):
+                    os.remove("test-{}".format(port))
 
 
 @pytest.fixture(scope="module")
-def blockchain_ursulas(three_agents):
+def blockchain_ursulas(three_agents, temp_config_root):
 
     token_agent, miner_agent, policy_agent = three_agents
     etherbase, alice, bob, *all_yall = token_agent.blockchain.interface.w3.eth.accounts
@@ -198,7 +226,8 @@ def blockchain_ursulas(three_agents):
 
     _ursulas = make_decentralized_ursulas(ether_addresses=ursula_addresses,
                                           miner_agent=miner_agent,
-                                          stake=True)
+                                          stake=True,
+                                          config_root=temp_config_root)
 
     try:
         yield _ursulas
